move authorization schema class
